+++ manual/scripts/extract_programs.sh
@@ -1,11 +1,19 @@
 #!/bin/bash -e
 
-VOTCASHARE="$(csg_call --show-share)"
+die() {
+  echo "$*" >&2
+  exit 1
+}
+[[ -z ${VOTCASHARE} ]] && die "${0##*/}: VOTCASHARE not set"
+export VOTCASHARE
 texfile=$1
+shift
 
-programs="ctp_map ctp_dump ctp_tools ctp_run ctp_parallel moo_overlap kmc_run"
+programs="ctp_map ctp_dump ctp_tools ctp_run ctp_parallel moo_overlap"
 rm -f $texfile; touch $texfile
 
 for program in $programs; do
-	$program --tex >> $texfile 
+	exe=$(echo "$@" | xargs -n1 echo | grep "$program")
+	echo ${exe}
+	$exe --tex >> $texfile 
 done
